requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/LLM
openai==1.12.0
anthropic==0.18.0
langchain==0.1.6
langchain-openai==0.0.5

# Document processing
pillow==10.2.0
pdf2image==1.17.0
PyPDF2==3.0.1
pytesseract==0.3.10
opencv-python==4.9.0.80

# Web automation
playwright==1.41.0
selenium==4.17.0

# API and UI
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0